♻️ refactor(db): merge connectorTool schema into connector.ts
